FeaGenerator - creating members

+++ examples/iom/csharp/ConnectionIomGenerator/ConnectionIomGenerator/Model/Member.cs
@@ -5,9 +5,9 @@
 		public int Id { get; set; }
 		public required string Name { get; set; }
 		public required string CrossSection { get; set; }
-		public double Direction { get; set; }
-		public double Pitch { get; set; }
-		public double Rotation { get; set; }
+		public float Direction { get; set; }
+		public float Pitch { get; set; }
+		public float Rotation { get; set; }
 		public bool IsContinuous { get; set; }
 	}
 }
